requirements.txt
lockfile
Django
django-crispy-forms
django-extensions
djangorestframework
Pillow
django-geoposition
pyyaml
ua-parser
user-agents
django-user-agents